refactor sheetService
